mcp-local/requirements.txt
torch
usearch
pyyaml
boto3
requests
mcp
sentence-transformers
fastmcp